svm: Allocate IOPM physically contiguously
The IOPM resides in 3 pages which _must_ be physically contiguous, as
they're referenced by the CPU.  Allocate them with contig_alloc(), which
we really shouldn't be using.

This appears to prevent both the rtls reset failure and the ATA
timeouts, as both were using I/O ports which would fall into the second
page of the IOPM which, when not physically contiguous, is essentially
random, meaning that a random selection of IO port accesses were never
intercepted.